:recycle:(trainer): extract NSGA-II from ga-runner.ts into dedicated module
- Move dominates, nondominatedSort*, assignCrowding, buildPopulationMeta
  and ObjectiveVector/PopulationMeta types to new nsga2.ts
- Replace duplicate estimateComplexity with shared import from
  complexity-estimator.ts
- ga-runner.ts reduced from 924 to ~640 lines

Refs #176
